+++ src/ast/stmt.rs
@@ -173,7 +173,8 @@ impl fmt::Debug for OpAssignment {
     }
 }
 
-/// An expression with a condition.
+/// _(internals)_ An expression with a condition.
+/// Exported under the `internals` feature only.
 ///
 /// The condition may simply be [`Expr::BoolConstant`] with `true` if there is actually no condition.
 #[derive(Debug, Clone, Default, Hash)]
@@ -1017,7 +1018,7 @@ impl Stmt {
     pub fn walk<'a>(
         &'a self,
         path: &mut Vec<ASTNode<'a>>,
-        on_node: &mut impl FnMut(&[ASTNode]) -> bool,
+        on_node: &mut (impl FnMut(&[ASTNode]) -> bool + ?Sized),
     ) -> bool {
         // Push the current node onto the path
         path.push(self.into());